CategoryTree renamed to FiltersTree, folders with FilterColors & FilterSizes replaced to FiltersTree directory
