+++ src/components/FiltersTree/FilterColors/FilterColorsStyles.js
