+++ src/components/FiltersTree/FilterSizes/FilterSizesTreeItem.js
@@ -2,9 +2,9 @@ import React from 'react';
 import {withStyles} from '@material-ui/core/styles';
 import {ThemeProvider} from '@material-ui/core';
 import theme from '../FilterColors/FilterColorsTheme';
-import TransitionComponent from '../TransitionComponent/TransitionComponent';
+import TransitionComponent from '../../TransitionComponent/TransitionComponent';
 import TreeItem from '@material-ui/lab/TreeItem';
-import { colors } from '../../styles/colorKit';
+import { colors } from '../../../styles/colorKit';
 
 const FilterColorsTreeItem = withStyles(() => ({
   iconContainer: {
