+++ src/components/StyledTreeItem/StyledTreeItem.js
@@ -1,7 +1,7 @@
 import {withStyles} from '@material-ui/core/styles';
 import {colors} from '../../styles/colorKit';
 import {ThemeProvider} from '@material-ui/core';
-import theme from '../CategoryTree/CategoryTreeTheme';
+import theme from '../FiltersTree/FiltersTreeTheme';
 import TransitionComponent from '../TransitionComponent/TransitionComponent';
 import TreeItem from '@material-ui/lab/TreeItem';
 import React from 'react';
